Move best_containing from Client to Backend
